Removed hardcoded paths. Ported to modern SDK. Added ARM64. Enabled bitcode support.

+++ freetype_ios.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 # Place this file in the same folder as a freetype SVN checkout.
-VERSION="6.1"
-MINVERSION="4.3"
-ARCHES=("armv7" "armv7s" "i386")
-PLATFORMS=("iPhoneOS" "iPhoneOS" "iPhoneSimulator")
+VERSION="9.1"		# SDK version
+MINVERSION="6.1"	# Target version.
+ARCHES=("arm64" "armv7" "armv7s" "i386")
+PLATFORMS=("iPhoneOS" "iPhoneOS" "iPhoneOS" "iPhoneSimulator")
 MERGE="lipo -create -output libfreetype-fat.a "
 
 for i in ${!ARCHES[@]} 
@@ -11,18 +11,28 @@ do
     # Create required paths:
     ARCH=${ARCHES[${i}]}
     PLATFORM=${PLATFORMS[${i}]}
-    SDK="/Developer/Platforms/${PLATFORM}.platform/Developer/SDKs/${PLATFORM}${VERSION}.sdk"
     SYSROOT="/Applications/Xcode.app/Contents/Developer/Platforms/${PLATFORM}.platform/Developer/SDKs/${PLATFORM}${VERSION}.sdk"
-    OUT="./libFreetype2-${ARCH}.a"
+    SDK=${SYSROOT}
+	OUT="./libFreetype2-${ARCH}.a"
     
+	if [ ! -e ${SDK} ]; then
+		echo "SDK not found at:"
+		echo ${SDK}
+		echo ""
+		echo "Try changing the 'version' inside this bash file. Here are some installed SDKs:"
+		echo $(ls $(dirname ${SDK}))
+		echo ""
+		exit 1
+	fi
+	
     # Clean old remnants:
     make clean; rm -rf objs/.libs/*
 
     # Usual configure and make routines:
-    ./configure --prefix=/Users/gerjo/freetype-2.4.10/build-${ARCH} --host=arm-apple-darwin --enable-static=yes --enable-shared=no \
-    CPPFLAGS="-arch ${ARCH} -fpascal-strings -Os -fmessage-length=0 -fvisibility=hidden -miphoneos-version-min=${VERSION} -I${SDK}/usr/include/libxml2 -isysroot ${SYSROOT}" \
+    ./configure --prefix=${PWD}/build-${ARCH} --host=arm-apple-darwin --enable-static=yes --enable-shared=no --with-bzip2=no \
+    CPPFLAGS="-arch ${ARCH} -fpascal-strings -Os -fembed-bitcode-marker -fembed-bitcode -fmessage-length=0 -fvisibility=hidden -miphoneos-version-min=${VERSION} -I${SDK}/usr/include/libxml2 -isysroot ${SYSROOT}" \
     CC=`xcrun -sdk iphoneos -find clang` \
-    CFLAGS="-arch ${ARCH} -fpascal-strings -Os -fmessage-length=0 -fvisibility=hidden -miphoneos-version-min=${VERSION} -isysroot ${SYSROOT}" \
+    CFLAGS="-arch ${ARCH} -fpascal-strings -Os -fembed-bitcode-marker -fembed-bitcode -fmessage-length=0 -fvisibility=hidden -miphoneos-version-min=${VERSION} -isysroot ${SYSROOT}" \
     LD=`xcrun -sdk iphoneos -find ld` \
     LDFLAGS="-arch ${ARCH} -isysroot ${SYSROOT} -miphoneos-version-min=${MINVERSION}" \
     AR=`xcrun -sdk iphoneos -find ar` && make
@@ -32,7 +42,13 @@ do
     MERGE="${MERGE} ${OUT}"
 done
 
+
+echo ""
 # Execute the merge command, output some debug info:
 echo `${MERGE}` && lipo -info libfreetype-fat.a
 
+echo ""
+echo "If there was an error, try running \"xcode-select --install\" (in particular for x86 simulator builds.";
+
+echo ""
 exit 0
